Initialize 'rc' so fine optimized code still works... Ugh.
re-order some conf file parsing for permissions reasons. Then take root
permissions back to open keys/limits files.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@781 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/endpoint.c
@@ -548,7 +548,7 @@ success:
 		/*
 		 * Don't make buffer larger than "default"
 		 */
-		size = MIN(size,statbuf.st_blksize);
+		size = MIN(size,(size_t)statbuf.st_blksize);
 
 		/*
 		 * buffer needs to be at least as large as one record.
@@ -1300,6 +1300,7 @@ recvfromttl(
 				cmdmsgptr->cmsg_type);
 		return -rc;
 NEXTCMSG:
+		;
 	}
 
 	return rc;
